+++ tests/ops/op_math/test_evolve.py
@@ -19,7 +19,7 @@ import pytest
 
 import pennylane as qml
 from pennylane.operation import AnyWires
-from pennylane.ops import Evolve, ParametrizedHamiltonian, QubitUnitary
+from pennylane.ops import Evolution, Evolve, ParametrizedHamiltonian, QubitUnitary
 
 
 class MyOp(qml.RX):  # pylint: disable=too-few-public-methods
@@ -240,3 +240,33 @@ class TestIntegration:
         assert qml.math.allclose(
             jax.grad(circuit)(params, t), jax.grad(true_circuit)(params, t), atol=1e-2
         )
+
+
+class TestEvolveConstructor:
+    """Unit tests for the evolve function"""
+
+    def test_evolve_returns_evolution_op(self):
+        """Test that the evolve function returns the `Evolution` operator when the input is
+        a generic operator."""
+        op = qml.s_prod(2, qml.PauliX(0))
+        final_op = qml.evolve(op)
+        assert isinstance(final_op, Evolution)
+
+    def test_matrix(self):
+        """Test that the matrix of the evolved function is correct."""
+        op = qml.s_prod(2, qml.PauliX(0))
+        final_op = qml.evolve(op)
+        mat = qml.math.expm(1j * qml.matrix(op))
+        assert qml.math.allequal(qml.matrix(final_op), mat)
+
+    def test_evolve_returns_callable(self):
+        """Test that the evolve function returns a callable when the input is a
+        ParametrizedHamiltonian."""
+        coeffs = [1, 2, 3]
+        ops = [qml.PauliX(0), qml.PauliY(1), qml.PauliZ(2)]
+        H = ParametrizedHamiltonian(coeffs=coeffs, observables=ops)
+        final_op = qml.evolve(H)
+        assert callable(final_op)
+        param_evolution = final_op(params=[], t=1)
+        assert isinstance(param_evolution, Evolve)
+        assert param_evolution.H is H
