+++ tests/tests_passing_pylint.txt
@@ -1,5 +1,4 @@
 tests/data/test_dataset.py
 tests/data/test_dataset_access.py
 tests/ops/op_math/test_parametrized_hamiltonian.py
-tests/ops/op_math/test_evolve_op.py
-tests/ops/functions/test_evolve.py
+tests/ops/op_math/test_evolve.py
